refactor(assistant): extract nested components from AssistantSuggestionCard
Extract `CatchMethodSection` and `PokemonListSection` from `AssistantSuggestionCard.tsx` into their own dedicated files. Replace `Object.entries` with the typed `objectEntries` utility.
