+++ .jules/sculptor.md
@@ -40,3 +40,7 @@ Extracted complex inline IIFE math calculations and tailwind styling logic for c
 - AI Readability Impact: Removes dense blocks of dynamic conditional rendering from the JSX structure, making the main functional component significantly easier to parse. This pattern cleans up the return block, separating domain-specific logic and UI flags from the presentation layer.
 - **AI Readability Refactoring Pattern**: Extracted complex inline mapping functions for rendering catch methods and generic suggestion lists into dedicated, well-named functional components (`CatchMethodSection` and `PokemonListSection`) in `AssistantSuggestionCard.tsx`.
 - AI Readability Impact: Removes dense blocks of dynamic conditional rendering and complex destructuring from the JSX structure, making the main functional component significantly easier to parse. This pattern cleans up the return block, separating domain-specific logic (like determining if a player has the correct rod for fishing encounters) from the presentation layer, thus reducing cognitive load and lowering AST parsing difficulty.
+- **AI Readability Refactoring Pattern**: Extracted complex nested functional components (`CatchMethodSection` and `PokemonListSection`) out of `AssistantSuggestionCard.tsx` into their own dedicated files.
+- AI Readability Impact: Significantly reduces the file size and AST complexity of `AssistantSuggestionCard.tsx`, making the primary presentation logic and conditional rendering structures much easier for AI to parse.
+- **AI Readability Refactoring Pattern**: Replaced `Object.entries` with the custom `objectEntries` utility from `src/utils/object.ts` in `AssistantSuggestionCard.tsx`.
+- AI Readability Impact: Provides strict type safety for tuples generated from mapped objects instead of relying on generic string keys, removing a common source of type uncertainty and cognitive load for code reasoning.
